# docs/api-spec.yml
openapi: 3.0.3
info:
  title: CinemaCore API
  description: >
    API Gateway specification for the CinemaCore Microservices ecosystem. 
    This acts as the single source of truth for frontend-backend contracts.
  version: 1.0.0
servers:
  - url: /api/v1
    description: API Gateway Entry Point
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProyeccionResponse:
      type: object
      properties:
        id:
          type: integer
        codigo:
          type: string
        descripcion:
          type: string
    SalaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        capacidad:
          type: integer
    FuncionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        movieId:
          type: integer
          format: int64
        sala:
          $ref: '#/components/schemas/SalaResponse'
        proyeccion:
          $ref: '#/components/schemas/ProyeccionResponse'
        fechaInicio:
          type: string
          format: date-time
        fechaFin:
          type: string
          format: date-time
        precioTicket:
          type: number
        status:
          type: string
paths:
...
  /showtimes:
    get:
      summary: (Público) Obtiene funciones disponibles por fecha
      operationId: getShowtimesByDate
      tags: [Showtimes]
      parameters:
        - name: fecha
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Lista de funciones programadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FuncionResponse'
  /showtimes/movies/active:
    get:
      summary: (Interno/BFF) Obtiene IDs de películas con funciones activas
      operationId: getActiveMovieIds
      tags: [Showtimes]
      responses:
        '200':
          description: Arreglo de IDs de películas con funciones futuras programadas
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
                  format: int64

    post:
      summary: (Admin) Programa una nueva función
      operationId: programarFuncion
      tags: [Showtimes]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                movie_id:
                  type: integer
                  format: int64
                sala_id:
                  type: integer
                  format: int64
                proyeccion_id:
                  type: integer
                fecha_inicio:
                  type: string
                  format: date-time
                duracion_min:
                  type: integer
                precio_ticket:
                  type: number
      responses:
        '201':
          description: Función programada y asientos pre-generados
        '409':
          description: Solapamiento de horarios en la misma sala
  /showtimes/admin:
    get:
      summary: (Admin) Consultar logística completa
      operationId: getAllShowtimesAdmin
      tags: [Showtimes]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: salaId
          in: query
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lista completa de funciones (sin filtros forzados)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FuncionResponse'
  /showtimes/{id}:
    delete:
      summary: (Admin) Cancelar función
      operationId: cancelShowtime
      tags: [Showtimes]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Función cancelada y asientos liberados/desactivados
...